Package cmdliner

Cmdliner provides a simple and compositional mechanism to convert command line arguments to OCaml values and pass them to your functions.

The library automatically handles syntax errors, help messages and UNIX man page generation. It supports programs with single or multiple commands (like git) and respect most of the POSIX and GNU conventions.

Manuals

The following manuals are available.

The tutorial gets you through the steps to write your first command line interface with Cmdliner.

The Command line interface manual describes how command lines and environment variables are parsed by Cmdliner.

Tool man pages describes how Cmdliner generates man pages for your tools and how you can format them.

The examples page has a few annoted examples that show to express the command line interface of a few classic tools with Cmdliner

API

Cmdliner Declarative definition of command line interfaces.

Package info

changes-files

CHANGES.md

license-files

LICENSE.md

readme-files

README.md