Package logs

Logs provides a logging infrastructure.

Logging is performed on sources whose reporting level can be set independently. Log message report is decoupled from logging and is handled by a reporter. A few optional log reporters are distributed with the package and the API easily allows to implement your own.

See the basics.

Library logs

Logs Logging.

Libraries logs.{browser,cli,fmt,lwt,threaded}

Each of these modules lives in its own library.

Logs_browser

Logs_cli Cmdliner support for Logs.

Logs_fmt Format colorful reporter for Logs.

Logs_lwt Lwt logging.

Logs_threaded Thread safe logging.

Package info

changes-files

CHANGES.md

license-files

LICENSE.md

readme-files

README.md