Package mtime

Mtime has platform independent support for monotonic wall-clock time. This time increases monotonically and is not subject to operating system calendar time adjustments. The library has types to represent nanosecond precision timestamps and time spans.

Mtime_clock provides acces to a monotonic system clock and its resolution (if available).

Library mtime

Mtime Monotonic time values.

Library mtime.clock.os

Mtime_clock Monotonic time clock.

This library also works with JavaScript, see the platform support.

Package info

changes-files

CHANGES.md

license-files

LICENSE.md

readme-files

README.md