Package uri

Uri Uniform Resource Identifier handling that is RFC3986-compliant.

Uri_services Module that maps the more common IANA well-known ports to and from their textual service names. To save memory, this module does not contains only the more common services. Use the Uri_services_full module if you need the complete list.

Uri_services_full Module that maps the IANA well-known ports to and from their textual service names

Package info

changes-files

CHANGES.md

license-files

LICENSE.md

readme-files

README.md